se agrega validacion en los servicios de promesas para que los operadores unicamente tengan acceso a sus promesas, en caso de que accedan a otra promesa que no les corresponda, el sistema tirara un error 403

+++ Backend/src/main/java/com/crmpps/Backend/repository/PromesaCustomRepository.java
@@ -24,7 +24,7 @@ public class PromesaCustomRepository {
             String tipoCumplimiento,
             LocalDate fechaCargaDesde,
             LocalDate fechaCargaHasta,
-            Integer operador,
+            Long operador,
             Boolean duplica
     ) {
         StringBuilder sb = new StringBuilder("SELECT a FROM promesa_pago a WHERE 1=1 ");
